refactor: reorganize project structure with better file names and test organization
## File Reorganization
- Keep `contract.ts` for NEAR Chain Signatures MPC contract
- Merge contract types into existing empty `types.ts` file
- Move all tests to `tests/` directory with clear structure:
  - `tests/unit/` for unit tests (omni-key.test.ts, mpc-contract.test.ts)
  - `tests/integration/` for integration tests (contract-signatures.test.ts, contract-view-methods.test.ts)

## Benefits
- **Cleaner structure**: All tests in `tests/` directory
- **Clear separation**: Unit vs integration tests
- **Simpler imports**: No dashes in filenames
- **Better organization**: Related types and schemas consolidated
- **Consistent patterns**: Follow standard project layout conventions

## Updated
- Package.json test scripts to use `tests/unit/` directory
- CI workflow to run `tests/unit/` for unit tests
- All import paths updated for new structure
- Index.ts exports cleaned up to avoid duplication
