+++ package.json
@@ -4,11 +4,11 @@
   "type": "module",
   "scripts": {
     "build": "tsc",
-    "test": "bun test tests/*.test.ts",
-    "test:unit": "bun test tests/*.test.ts",
+    "test": "bun test tests/unit/",
+    "test:unit": "bun test tests/unit/",
     "test:integration": "bun test tests/integration/",
     "test:all": "bun test",
-    "test:watch": "bun test --watch tests/*.test.ts",
+    "test:watch": "bun test --watch tests/unit/",
     "test:ci": "bun test --coverage",
     "lint": "biome check --write .",
     "typecheck": "tsc --noEmit",
